Working room xyz sliders and automatic player positioning. Going to extend the latter to the source and add the source panel from the previous implementation.
